+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS37FSSEPT3THEG7NG2E18R6"
+        "01KS4G97SEMADQNWPSGKXDWJ8T"
     )
     r"""The unique identifier of the data source"""