+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS37FSSD02NCQKX37P99C2FM"
+        "01KS4G97SDMBSAT43CDAVYF014"
     )
     r"""The unique identifier of the data source"""