+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KS37FSSG9125TJ07RM61BJG8"
+        "01KS4G97SHR6MPC27Y95W9Y1QJ"
     )
     r"""The unique identifier of the data source"""